# app/backend/docs/swagger.yaml
definitions:
...
  dto.CreateSemesterDto:
    properties:
      semester:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - semester
    type: object
...
  dto.UpdateSemesterDto:
    properties:
      semester:
        maxLength: 255
        minLength: 3
        type: string
    type: object
...
  httperr.Fields:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
  httperr.RestErr:
    properties:
      code:
        type: integer
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/httperr.Fields'
        type: array
      message:
        type: string
    type: object
...
  response.ManySemestersResponse:
    properties:
      semesters:
        items:
          $ref: '#/definitions/response.SemesterResponse'
        type: array
    type: object
...
  response.SemesterResponse:
    properties:
      semester:
        type: string
      uuid:
        type: string
    type: object
...
info:
  contact:
    email: your@mail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Time Table API
  version: "1.0"
paths:
...
  /semesters:
    post:
      consumes:
      - application/json
      description: Endpoint for create semester
      parameters:
      - description: Create semester dto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSemesterDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Create new semester
      tags:
      - semester
  /semesters/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete semester by uuid
      parameters:
      - description: semester uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Delete semester
      tags:
      - semester
    get:
      consumes:
      - application/json
      description: Get semester by uuid
      parameters:
      - description: semester uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SemesterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Semester details
      tags:
      - semester
    patch:
      consumes:
      - application/json
      description: Endpoint for update semester
      parameters:
      - description: semester uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: Update semester dto
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.UpdateSemesterDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Update semester
      tags:
      - semester
  /semesters/list-all:
    get:
      consumes:
      - application/json
      description: Get many semesters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ManySemestersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Get many semesters
      tags:
      - semester
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"