# app/backend/internal/handler/response/semester_response.go
package response

type SemesterResponse struct {
	UUID     string `json:"uuid"`
	Semester string `json:"semester"`
}

type ManySemestersResponse struct {
	Semesters []SemesterResponse `json:"semesters"`
}